chore(refact): move paginate_records fn outside of datasetio
